Extract menu footer to separate file
